refactor: seperate each payment result state is seperated composable
